rooms: create geometry module
This moves geometry related room functions to their own module.
