+++ src/libtrx/include/libtrx/game/rooms.h
@@ -5,4 +5,5 @@
 #include "rooms/draw.h"
 #include "rooms/enum.h"
 #include "rooms/floor_data.h"
+#include "rooms/geometry.h"
 #include "rooms/utils.h"
