+++ src/libtrx/include/libtrx/game/rooms/common.h
@@ -19,8 +19,6 @@ void Room_IncrementFlipTimer(int32_t num_frames);
 int32_t Room_GetFlipSlotFlags(int32_t slot_idx);
 void Room_SetFlipSlotFlags(int32_t slot_idx, int32_t flags);
 
-extern void Room_AlterFloorHeight(const ITEM *item, int32_t height);
-
 int32_t Room_GetAdjoiningRooms(
     int16_t init_room_num, int16_t out_room_nums[], int32_t max_room_num_count);
 
@@ -28,32 +26,6 @@ int16_t Room_GetIndexFromPos(int32_t x, int32_t y, int32_t z);
 int32_t Room_FindByPos(int32_t x, int32_t y, int32_t z);
 int32_t Room_GetFlippedBaseRoom(int32_t room_num);
 BOUNDS_32 Room_GetWorldBounds(void);
-
-extern SECTOR *Room_GetSector(
-    int32_t x, int32_t y, int32_t z, int16_t *room_num);
-SECTOR *Room_GetWorldSector(const ROOM *room, int32_t x_pos, int32_t z_pos);
-SECTOR *Room_GetUnitSector(
-    const ROOM *room, int32_t x_sector, int32_t z_sector);
-SECTOR *Room_GetPitSector(const SECTOR *sector, int32_t x, int32_t z);
-SECTOR *Room_GetSkySector(const SECTOR *sector, int32_t x, int32_t z);
-
-void Room_SetAbyssHeight(int16_t height);
-bool Room_IsAbyssHeight(int32_t height);
-HEIGHT_TYPE Room_GetHeightType(void);
-int16_t Room_GetHeight(const SECTOR *sector, int32_t x, int32_t y, int32_t z);
-int16_t Room_GetHeightEx(
-    const SECTOR *sector, int32_t x, int32_t y, int32_t z, bool fix_tilts);
-extern int32_t Room_GetWaterHeight(
-    int32_t x, int32_t y, int32_t z, int16_t room_num);
-int16_t Room_GetCeiling(const SECTOR *sector, int32_t x, int32_t y, int32_t z);
-int16_t Room_GetCeilingEx(
-    const SECTOR *sector, int32_t x, int32_t y, int32_t z, bool fix_tilts);
-extern int16_t Room_GetTiltType(
-    const SECTOR *sector, int32_t x, int32_t y, int32_t z);
-extern int32_t Room_FindGridShift(int32_t src, int32_t dst);
-
-bool Room_IsOnWalkable(
-    const SECTOR *sector, int32_t x, int32_t y, int32_t z, int32_t room_height);
 
 extern void Room_GetNearbyRooms(
     int32_t x, int32_t y, int32_t z, int32_t r, int32_t h, int16_t room_num);
